refactorer: clean up vehicle reservation release/module placement (#550)
Drop the redundant requiredVehicleRole guard around releaseVehicleReservation
in TaskDispatch.ts's cancelAction/interruptActiveAction — the function
already no-ops when nothing is reserved for the action id. Move
findVehicleForClaim and promoteVehicleGatedAction from GameLoop.ts into
VehicleReservation.ts, which already owns every other vehicle-reservation
transition; verified no new import-cycle SCC membership beyond the
already-cyclic engine/entities core. Expand the tick-reordering comment in
console/commands/events.ts to state plainly that moving the completion pass
before the claim pass affects every task completion, not just vehicle-gated
ones.

Declined moving ActionSelection.ts's seedTaskTimerFields into TaskDispatch.ts
(suggestion 3) — doing so pulls ActionSelection.ts (and its nav/Pathfinding.ts
dependency) into the same already-existing engine-wide circular-dependency
component via a new path, which the module's own header comment was written
to avoid; left as-is per the task's own guidance to skip if not clean.
