+++ src/core/engine/VehicleReservation.ts
@@ -1,21 +1,24 @@
 // BlastSimulator2026 — Vehicle reservation for vehicle-gated actions (#550)
-// Owns the exclusive claim a PendingAction holds on a Vehicle from the
-// moment an employee claims a vehicle-gated action until the action
-// completes, is cancelled, or the vehicle is destroyed underneath it.
-// Core-pure: imports only from entities/, state/, and EntityMovementTick.js
-// (one-way, EntityMovementTick.ts never imports back). Deliberately does NOT
-// import TaskDispatch.ts — TaskDispatch.ts imports releaseVehicleReservation
-// from here, so calling into it back would be a cycle. Where reconciliation
-// needs to interrupt an active action (case (c) below), it reports the need
-// to its caller instead of performing the interruption itself — see
-// reconcileVehicleReservations's return type and ArrivalGate.ts, the sole
-// caller, which already imports both modules safely.
+// Owns the exclusive claim a PendingAction holds on a Vehicle: finding one to
+// claim, reserving it, promoting the claim to an active walk-to-vehicle
+// transition, and releasing it — whether that's on completion, cancellation,
+// needs-interruption, or the vehicle being destroyed underneath it.
+// GameLoop.ts's claim/promotion sites call into this module rather than
+// duplicating any of it. Core-pure: imports only from entities/, state/, and
+// EntityMovementTick.js (one-way, EntityMovementTick.ts never imports back).
+// Deliberately does NOT import TaskDispatch.ts — TaskDispatch.ts imports
+// releaseVehicleReservation from here, so calling into it back would be a
+// cycle. Where reconciliation needs to interrupt an active action (case (c)
+// below), it reports the need to its caller instead of performing the
+// interruption itself — see reconcileVehicleReservations's return type and
+// ArrivalGate.ts, the sole caller, which already imports both modules safely.
 
-import type { GameState } from '../state/GameState.js';
+import type { GameState, PendingAction } from '../state/GameState.js';
 import type { Employee } from '../entities/Employee.js';
 import type { Vehicle, VehicleRole } from '../entities/Vehicle.js';
-import { unassignDriver } from '../entities/Vehicle.js';
+import { unassignDriver, moveVehicle } from '../entities/Vehicle.js';
 import { ROLE_LICENCE_REQUIRED } from '../entities/VehicleDriverAssignment.js';
+import { requestBoardVehicle } from '../entities/VehicleBoarding.js';
 import { setVehicleIdle } from './EntityMovementTick.js';
 
 /** True when `employee` holds the licence a vehicle of `role` requires (ROLE_LICENCE_REQUIRED, VehicleDriverAssignment.ts). */
@@ -53,6 +56,63 @@ export function findFreeVehicleForRole(state: GameState, role: VehicleRole, empl
 /** Marks `vehicle` reserved for `actionId`. Caller must have already confirmed the vehicle came from findFreeVehicleForRole this same tick. */
 export function reserveVehicle(vehicle: Vehicle, actionId: number): void {
   vehicle.reservedForActionId = actionId;
+}
+
+/**
+ * Vehicle-gate check shared by both of GameLoop.ts's claim sites (#550): for
+ * a non-vehicle action this is always a pass-through no-op; for a
+ * vehicle-gated one it finds (but does not yet reserve) a qualifying free
+ * vehicle via findFreeVehicleForRole above. `ok: false` means this employee
+ * cannot claim this action right now — same "stays queued, retries next
+ * tick" outcome as selectBestActionForEmployee returning null for an
+ * unreachable target, not an error.
+ */
+export function findVehicleForClaim(
+  state: GameState,
+  action: PendingAction,
+  employee: Employee,
+): { ok: true; vehicle: Vehicle | null } | { ok: false } {
+  if (action.requiredVehicleRole === null) return { ok: true, vehicle: null };
+  const vehicle = findFreeVehicleForRole(state, action.requiredVehicleRole, employee);
+  return vehicle === null ? { ok: false } : { ok: true, vehicle };
+}
+
+/**
+ * Vehicle-gated claim transition (#550): instead of walking to the action's
+ * own target, the employee walks to (or, already driving it — the
+ * continuity tie-break in findFreeVehicleForRole above — skips straight
+ * past) the vehicle reserved for this action at claim time. Called by
+ * GameLoop.ts's promoteActionToActive when action.requiredVehicleRole is set.
+ *
+ * Deliberately does not call seedTaskTimerFields here — that only happens
+ * once the VEHICLE (not the employee) reaches action.targetX/targetZ, via
+ * the vehicle-drive loop ArrivalGate.tickArrivalGate adds. Boarding and
+ * driving reuse existing machinery end to end: VehicleBoarding.requestBoardVehicle
+ * for the walk-and-board (resolved by ArrivalGate.resolveBoarding), and
+ * Vehicle.moveVehicle for setting the vehicle's own destination once someone
+ * is already aboard — EntityMovementTick.tickVehicle is the only thing that
+ * ever actually advances a vehicle's x/z (see ArrivalGate.ts's header).
+ */
+export function promoteVehicleGatedAction(state: GameState, employee: Employee, action: PendingAction): void {
+  const vehicle = state.vehicles.vehicles.find(v => v.reservedForActionId === action.id);
+  // Reservation vanished between claim and promotion (shouldn't happen within
+  // a single tick, but reconcileVehicleReservations is the backstop if it
+  // ever does) — leave the employee idle-but-claimed; next tick's reconcile
+  // sweep interrupts the action back to the pool.
+  if (!vehicle) return;
+
+  if (vehicle.driverId === employee.id) {
+    moveVehicle(state.vehicles, vehicle.id, action.targetX, action.targetZ);
+    return;
+  }
+
+  // Stage the destination now (the same targetX/targetZ tickVehicle reads),
+  // but deliberately leave vehicle.task alone — starting it 'moving' before
+  // anyone is aboard would drive an unmanned vehicle. ArrivalGate.resolveBoarding
+  // flips it to 'moving' once the employee has actually boarded.
+  vehicle.targetX = action.targetX;
+  vehicle.targetZ = action.targetZ;
+  requestBoardVehicle(state, vehicle.id, employee.id);
 }
 
 /**
